+++ CHANGELOG.md
@@ -8,4 +8,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 
-…
+- Module service provider
+- Base Eloquent model
+- Queueable job